Revert "Convert float to double"
This reverts commit 448b3ad42c883155c9e9ffdd99f3e2c66c13f825.

+++ form_res_rels_jg.c
@@ -33,7 +33,7 @@ static int comp_rank_judged (), comp_sim_docno (), comp_docno ();
 /* Definitions used for temporary and cached values */
 typedef struct {
     char *docno;
-    double sim;
+    float sim;
     long rank;
     long rel;
 } DOCNO_INFO;
